Integrated code by David Boike from http://code.google.com/p/apns-sharp/source/detail?r=27#
Upgraded Notification to support the enhanced notification format described in
Apple developer documentation which includes an expiry and a (so-far) unused
identifier.

Also updated the building of the notification payload so that it was easier to
verify that the byte array was being constructed properly with no math errors.

http://developer.apple.com/iphone/library/documentation/NetworkingInternet/Conce
ptual/RemoteNotificationsPG/CommunicatingWIthAPS/CommunicatingWIthAPS.html